add adjust SavingMobileStat
