+++ src/pages/gd/Stake/Savings/index.tsx
@@ -11,7 +11,8 @@ import styled from 'styled-components'
 import AsyncTokenIcon from 'components/gd/sushi/AsyncTokenIcon'
 import useSendAnalyticsData from 'hooks/useSendAnalyticsData'
 import { useWindowSize } from 'hooks/useWindowSize'
-import { SavingsDepositMobile, SavingsMobileStat } from './SavingsDepositMobile'
+import { SavingsDepositMobile } from './SavingsDepositMobile'
+import { SavingsMobileStat } from './SavingsMobileStat'
 import { HeadingCopy } from 'components/Savings/SavingsCard'
 import { useActiveWeb3React } from 'hooks/useActiveWeb3React'
 import { ModalButton } from 'components/Savings/SavingsModal/ModalButton'
